add AJAX methods, rewrite spec for js, add make_better method

+++ app/controllers/answers_controller.rb
@@ -1,37 +1,27 @@
 class AnswersController < ApplicationController
-  before_action :authenticate_user!, except: %i[index show]
+  before_action :authenticate_user!, except: %i[index]
 
-  def show; end
-
-  def new; end
-
-  def edit; end
+  def index
+    question.answers
+  end
 
   def create
     @answer = question.answers.new(answer_params)
     @answer.user = current_user
-    if @answer.save
-      redirect_to @answer.question, notice: 'Your answer has been successfully added.'
-    else
-      render 'questions/show'
-    end
+    @answer.save
   end
 
   def update
-    if answer.update(answer_params)
-      redirect_to @answer.question, notice: 'Your answer successful updated!'
-    else
-      render :edit, alert: 'Your answer has not been saved!'
-    end
+    answer.update(answer_params)
   end
 
   def destroy
-    if current_user.owner?(answer)
-      answer.destroy
-      redirect_to answer.question, notice: 'Your answer successfully deleted.'
-    else
-      redirect_to answer.question, alert: "You can't delete not your answer!"
-    end
+    answer.destroy if current_user.owner?(answer)
+  end
+
+  def make_better
+    @question = answer.question
+    answer.up_to_best! if current_user.owner?(answer.question)
   end
 
   private
